requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
open3d>=0.17.0
trimesh>=3.21.0
pyshtools>=4.10.0
pyvista>=0.39.0
matplotlib>=3.7.0
pyyaml>=6.0
tqdm>=4.65.0
Pillow>=9.5.0
astropy>=5.3.0
